server/python-api/requirements.txt
# AI简历助手 - Python API依赖
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.7
httpx==0.27.0
pydantic==2.6.1
openai==1.13.3
docx==0.2.4
python-docx==1.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
langchain==0.1.9
tiktoken==0.6.0
numpy==1.26.4
